rename exercise and initial-assessment folders
